+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [next]
+ - Adding Game#pauseEngine and Game#resumeEngine methods
  - Removing FlameBinding since it isn't used and clashes with newest flutter
 
 ## 0.19.0